Struct shenlong_core::sierra::llvm_compiler::Compiler

pub struct Compiler<'a, 'ctx> {
    pub program: &'a Program,
    pub context: &'ctx Context,
    pub builder: &'a Builder<'ctx>,
    pub module: Module<'ctx>,
    pub variables: HashMap<String, BasicValueEnum<'ctx>>,
    pub output_path: String,
    pub state: CompilationState,
    pub valid_state_transitions: HashMap<CompilationStateTransition, bool>,
    pub types: HashMap<String, Box<dyn BasicType<'ctx> + 'a>>,
    pub id_from_name: HashMap<String, String>,
    pub main_calls: Vec<BasicValueEnum<'ctx>>,
}

Compiler is the main entry point for the LLVM backend. It is responsible for compiling a Sierra program to LLVM IR.

Fields
program: &'a Program

The Sierra program.

context: &'ctx Context

The LLVM context.

builder: &'a Builder<'ctx>

The LLVM builder.

module: Module<'ctx>

The LLVM module.

variables: HashMap<String, BasicValueEnum<'ctx>>

The variables of the program.

output_path: String

The output path.

state: CompilationState

The current compilation state.

valid_state_transitions: HashMap<CompilationStateTransition, bool>

The valid state transitions.

types: HashMap<String, Box<dyn BasicType<'ctx> + 'a>>

The types.

id_from_name: HashMap<String, String>

Mapping from type name to program id.

main_calls: Vec<BasicValueEnum<'ctx>>

Calls in the main function.

Implementations

impl<'a, 'ctx> Compiler<'a, 'ctx>

pub fn felt_add(
&mut self,
libfunc_declaration: &LibfuncDeclaration
) -> CompilerResult<()>

Implementation of the LLVM IR conversion of a felt addition.

Arguments

libfunc_declaration - The corelib function declaration of felt_add.

Error

Returns an error if the felt type has not been declared previously.

impl<'a, 'ctx> Compiler<'a, 'ctx>

pub fn felt_const(
&mut self,
libfunc_declaration: &LibfuncDeclaration
) -> CompilerResult<()>

Implementation of the LLVM IR conversion of a felt constant.

Arguments

libfunc_declaration - The corelib function declaration of felt_const.

Error

Returns an error if the felt type has not been declared previously.

impl<'a, 'ctx> Compiler<'a, 'ctx>

pub fn felt_sub(
&mut self,
libfunc_declaration: &LibfuncDeclaration
) -> CompilerResult<()>

Implementation of the LLVM IR conversion of a felt substraction.

Arguments

libfunc_declaration - The corelib function declaration of felt_sub.

Error

Returns an error if the felt type has not been declared previously.

impl<'a, 'ctx> Compiler<'a, 'ctx>

pub fn dup(
&mut self,
libfunc_declaration: &LibfuncDeclaration
) -> CompilerResult<()>

Implementation of the LLVM IR conversion of a duplication operation.

Arguments

libfunc_declaration - The corelib function declaration of dup.

Error

Returns an error if the type T has not been declared previously.

impl<'a, 'ctx> Compiler<'a, 'ctx>

pub fn rename(
&mut self,
libfunc_declaration: &LibfuncDeclaration
) -> CompilerResult<()>

Implementation of the LLVM IR conversion of the rename function.

Arguments

libfunc_declaration - The corelib function declaration of rename.

Error

Returns an error if the type T has not been declared previously.

impl<'a, 'ctx> Compiler<'a, 'ctx>

pub fn store_temp(
&mut self,
libfunc_declaration: &LibfuncDeclaration
) -> CompilerResult<()>

Implementation of the LLVM IR conversion of a store_temp operation.

Arguments

libfunc_declaration - The corelib function declaration of store_temp.

Error

Returns an error if the type T has not been declared previously.

impl<'a, 'ctx> Compiler<'a, 'ctx>

pub fn struct_construct(&mut self, libfunc_declaration: &LibfuncDeclaration)

Implementation of the LLVM IR conversion of a struct construction operation.

Arguments

libfunc_declaration - The corelib function declaration of StructConstruct.

Error

Returns an error if the type T has not been declared previously.

impl<'a, 'ctx> Compiler<'a, 'ctx>

pub fn struct_deconstruct(&mut self, libfunc_declaration: &LibfuncDeclaration)

Implementation of the LLVM IR conversion of a struct deconstruction operation.

Arguments

libfunc_declaration - The corelib function declaration of StructDeconstruct.

Error

Returns an error if the type T has not been declared previously.

impl<'a, 'ctx> Compiler<'a, 'ctx>

Implementation of the compiler.

pub fn compile_from_file(
program_path: &str,
output_path: &str
) -> CompilerResult<()>

Compile a Sierra program file to LLVM IR.

Arguments

program_path - The Sierra program to compile.

output_path - The path to the output LLVM IR file.

Returns

The result of the compilation.

Errors

If the compilation fails.

pub fn compile_from_code(
sierra_code: &str,
output_path: &str
) -> CompilerResult<()>

Compile a Sierra program code to LLVM IR.

Arguments

sierra_code - The Sierra program code.

output_path - The path to the output LLVM IR file.

Returns

The result of the compilation.

Errors

If the compilation fails.

pub fn compile_sierra_program_to_llvm(
program: Program,
output_path: &str
) -> CompilerResult<()>

Compiles a Sierra Program representation to LLVM IR.

Process overview

Create an LLVM context, builder and module.

Instantiate variables map.

Process the program type declarations.

Process the core library functions.

Process the program statements.

Finalize compilation and write the LLVM IR to a file.

Arguments

program - The Sierra program to compile.

output_path - The path to the output LLVM IR file.

Returns

The result of the compilation.

Errors

If the compilation fails.

Example

use std::fs;

use cairo_lang_sierra::ProgramParser;
use shenlong_core::sierra::llvm_compiler::Compiler;

let sierra_program_path = "../examples/program.sierra";
let llvm_ir_path = "../examples/program.ll";

// TODO: Find a way to make doc tests pass.
// Read the program from the file.
let sierra_code = fs::read_to_string(sierra_program_path).unwrap();
// Parse the program.
let program = ProgramParser::new().parse(&sierra_code).unwrap();
// Compile the program to LLVM IR.
let result = Compiler::compile_from_file(sierra_program_path, llvm_ir_path);
// Check the result.

pub fn check_state(
&self,
expected_state: &CompilationState
) -> CompilerResult<()>

Check if the compilation is in a valid state. If the compilation is not in a valid state, return an error.

Arguments

*expected_state - State the compiler should be in.

Returns

Result if the state is in an existing state.

Errors

If the compiler is in a state that doesn’t exist

pub fn move_to(&mut self, state: CompilationState) -> CompilerResult<()>

Move the compilation state to the next state.

Arguments

state - The new compilation state.

Errors

If the transition is not valid, return an error.

pub fn state(&self) -> &CompilationState

Get the current compilation state.

impl<'a, 'ctx> Compiler<'a, 'ctx>

Implementation of the corelib functions processing for the compiler.

pub fn process_core_lib_functions(&mut self) -> CompilerResult<()>

Process core library functions in the Sierra program.

Errors

if the processing of the core lib functions fails.

impl<'a, 'ctx> Compiler<'a, 'ctx>

Implementation for the type processing for the compiler.

pub fn process_funcs(&mut self) -> CompilerResult<()>

Process types in the Sierra program. For each type declaration in the Sierra program, create a corresponding type in the LLVM context.

Errors

If the processing of the sierra types fails.

impl<'a, 'ctx> Compiler<'a, 'ctx>

Implementation of the statement processing for the compiler.

pub fn process_statements_from_until(
&mut self,
from: usize,
until: Option<usize>
) -> CompilerResult<()>

Process statements in the Sierra program.

Errors

If the processing of the sierra statements fails.

impl<'a, 'ctx> Compiler<'a, 'ctx>

Implementation for the type processing for the compiler.

pub fn process_types(&mut self) -> CompilerResult<()>

Process types in the Sierra program. For each type declaration in the Sierra program, create a corresponding type in the LLVM context.

Errors

If the processing of the sierra types fails.

pub fn get_type_from_name(
&self,
type_name: &str
) -> CompilerResult<BasicTypeEnum<'ctx>>

Returns the LLVM IR type from the type name

Arguments

type_name - The type name (ex: “felt”)

Returns

The LLVM IR type.

impl<'a, 'ctx> Compiler<'a, 'ctx>

pub fn felt_is_zero(
&mut self,
invocation: &Invocation,
invocation_nb: usize
) -> CompilerResult<()>

Implementation of the LLVM IR conversion of felt == 0.

Arguments

invocation - The invocation object.

invocation_nb - The invocation number.

Error

Returns an error if the processing of the branches statements fails.

impl<'a, 'ctx> Compiler<'a, 'ctx>

pub fn felt(&mut self, type_declaration: &TypeDeclaration)

impl<'a, 'ctx> Compiler<'a, 'ctx>

pub fn non_zero(&mut self, type_declaration: &TypeDeclaration)

impl<'a, 'ctx> Compiler<'a, 'ctx>

pub fn sierra_struct(&mut self, type_declaration: &TypeDeclaration)