Struct shenlong_core::sierra::llvm_compiler::Compiler

pub struct Compiler<'a, 'ctx> {
    pub program: &'a Program,
    pub context: &'ctx Context,
    pub builder: &'a Builder<'ctx>,
    pub module: &'a Module<'ctx>,
    pub variables: HashMap<String, Option<PointerValue<'ctx>>>,
    pub output_path: &'a str,
    pub state: CompilationState,
    pub valid_state_transitions: HashMap<CompilationStateTransition, bool>,
}

Compiler is the main entry point for the LLVM backend. It is responsible for compiling a Sierra program to LLVM IR.

Fields
program: &'a Programcontext: &'ctx Contextbuilder: &'a Builder<'ctx>module: &'a Module<'ctx>variables: HashMap<String, Option<PointerValue<'ctx>>>output_path: &'a strstate: CompilationStatevalid_state_transitions: HashMap<CompilationStateTransition, bool>
Implementations

impl<'a, 'ctx> Compiler<'a, 'ctx>

Implementation of the compiler.

pub fn compile_from_file(program_path: &str, output_path: &str) -> Result<()>

Compile a Sierra program file to LLVM IR.

Arguments

program_path - The Sierra program to compile.

output_path - The path to the output LLVM IR file.

Returns

The result of the compilation.

Errors

If the compilation fails.

pub fn compile_from_code(sierra_code: &str, output_path: &str) -> Result<()>

Compile a Sierra program code to LLVM IR.

Arguments

sierra_code - The Sierra program code.

output_path - The path to the output LLVM IR file.

Returns

The result of the compilation.

Errors

If the compilation fails.

pub fn compile_sierra_program_to_llvm(
program: Program,
output_path: &str
) -> Result<()>

Compiles a Sierra Program representation to LLVM IR.

Process overview

Create an LLVM context, builder and module.

Instantiate variables map.

Process the program type declarations.

Process the core library functions.

Process the program statements.

Finalize compilation and write the LLVM IR to a file.

Arguments

program - The Sierra program to compile.

output_path - The path to the output LLVM IR file.

Returns

The result of the compilation.

Errors

If the compilation fails.

Example

use sierra::ProgramParser;
use shenlong_core::sierra::llvm_compiler::Compiler;
use std::fs;

let sierra_program_path = "examples/program.sierra";
let llvm_ir_path = "examples/program.ll";

// TODO: Find a way to make doc tests pass.
// Read the program from the file.
let sierra_code = fs::read_to_string(sierra_program_path).unwrap();
// Parse the program.
let program = ProgramParser::new().parse(&sierra_code).unwrap();
//Compile the program to LLVM IR.
let result = Compiler::compile_from_file(sierra_program_path, llvm_ir_path);
// Check the result.

pub fn state(&self) -> &CompilationState

Get the current compilation state.