Struct shenlong_core::sierra::llvm_compiler::Compiler

pub struct Compiler<'a, 'ctx> {
    pub program: &'a Program,
    pub context: &'ctx Context,
    pub builder: &'a Builder<'ctx>,
    pub module: Module<'ctx>,
    pub variables: HashMap<String, Option<PointerValue<'ctx>>>,
    pub output_path: String,
    pub state: CompilationState,
    pub valid_state_transitions: HashMap<CompilationStateTransition, bool>,
    pub types: HashMap<&'ctx str, Box<dyn BasicType<'ctx> + 'a>>,
    pub libfunc_processors: HashMap<String, Func<'ctx>>,
    pub main_calls: Vec<BasicValueEnum<'ctx>>,
}

Compiler is the main entry point for the LLVM backend. It is responsible for compiling a Sierra program to LLVM IR.

Fields
program: &'a Program

The Sierra program.

context: &'ctx Context

The LLVM context.

builder: &'a Builder<'ctx>

The LLVM builder.

module: Module<'ctx>

The LLVM module.

variables: HashMap<String, Option<PointerValue<'ctx>>>

The variables of the program.

output_path: String

The output path.

state: CompilationState

The current compilation state.

valid_state_transitions: HashMap<CompilationStateTransition, bool>

The valid state transitions.

types: HashMap<&'ctx str, Box<dyn BasicType<'ctx> + 'a>>

The types.

libfunc_processors: HashMap<String, Func<'ctx>>

The library functions processors. Each processor is responsible for processing a specific libfunc and generating the corresponding LLVM IR.

main_calls: Vec<BasicValueEnum<'ctx>>
Implementations

impl<'a, 'ctx> Compiler<'a, 'ctx>

Implementation of the compiler.

pub fn compile_from_file(
program_path: &str,
output_path: &str
) -> Result<(), CompilerError>

Compile a Sierra program file to LLVM IR.

Arguments

program_path - The Sierra program to compile.

output_path - The path to the output LLVM IR file.

Returns

The result of the compilation.

Errors

If the compilation fails.

pub fn compile_from_code(
sierra_code: &str,
output_path: &str
) -> Result<(), CompilerError>

Compile a Sierra program code to LLVM IR.

Arguments

sierra_code - The Sierra program code.

output_path - The path to the output LLVM IR file.

Returns

The result of the compilation.

Errors

If the compilation fails.

pub fn compile_sierra_program_to_llvm(
program: Program,
output_path: &str
) -> Result<(), CompilerError>

Compiles a Sierra Program representation to LLVM IR.

Process overview

Create an LLVM context, builder and module.

Instantiate variables map.

Process the program type declarations.

Process the core library functions.

Process the program statements.

Finalize compilation and write the LLVM IR to a file.

Arguments

program - The Sierra program to compile.

output_path - The path to the output LLVM IR file.

Returns

The result of the compilation.

Errors

If the compilation fails.

Example

use std::fs;

use cairo_lang_sierra::ProgramParser;
use shenlong_core::sierra::llvm_compiler::Compiler;

let sierra_program_path = "../examples/program.sierra";
let llvm_ir_path = "../examples/program.ll";

// TODO: Find a way to make doc tests pass.
// Read the program from the file.
let sierra_code = fs::read_to_string(sierra_program_path).unwrap();
// Parse the program.
let program = ProgramParser::new().parse(&sierra_code).unwrap();
// Compile the program to LLVM IR.
let result = Compiler::compile_from_file(sierra_program_path, llvm_ir_path);
// Check the result.

pub fn check_state(
&self,
expected_state: &CompilationState
) -> Result<(), CompilerError>

Check if the compilation is in a valid state. If the compilation is not in a valid state, return an error.

Arguments

*expected_state - State the compiler should be in.

Returns

Result if the state is in an existing state.

Errors

If the compiler is in a state that doesn’t exist

pub fn move_to(&mut self, state: CompilationState) -> Result<(), CompilerError>

Move the compilation state to the next state.

Arguments

state - The new compilation state.

Errors

If the transition is not valid, return an error.

pub fn state(&self) -> &CompilationState

Get the current compilation state.

impl<'a, 'ctx> Compiler<'a, 'ctx>

Implementation of the constant processing for the compiler.

pub fn process_const(
&mut self,
libfunc_declaration: &LibfuncDeclaration
) -> Result<String, CompilerError>

Process the constants of this sierra program.

Arguments

*libfunc_declaration - Declaration of the corelib function in the sierra file.

Returns

The constant mangled name.

Errors

If the processing of the consant fails.

impl<'a, 'ctx> Compiler<'a, 'ctx>

Implementation of the corelib functions processing for the compiler.

pub fn prepare_libfunc_processors(&mut self) -> Result<(), CompilerError>

Prepare the libfunc core processors (those are functions from the core lib).

Errors

If the preparation of the processing of the core lib functions fails.

pub fn process_core_lib_functions(&mut self) -> Result<(), CompilerError>

Process core library functions in the Sierra program.

Errors

if the processing of the core lib functions fails.

impl<'a, 'ctx> Compiler<'a, 'ctx>

Implementation of the statement processing for the compiler.

pub fn process_statements(&mut self) -> Result<(), CompilerError>

Process statements in the Sierra program.

Errors

If the processing of the sierra statements fails.

impl<'a, 'ctx> Compiler<'a, 'ctx>

Implementation for the type processing for the compiler.

pub fn process_types(&mut self) -> Result<(), CompilerError>

Process types in the Sierra program. For each type declaration in the Sierra program, create a corresponding type in the LLVM context.

Errors

If the processing of the sierra types fails.